some updates in the last day of 2020
4:00 am, Dec 31th, 2020
Last day of 2020

Still can't believe it has been the last day of 2020, it's bad, it's really really bad. I thought I could hang out with my friends for this summer, thought I could go to travel a few places that I have already thought about for a long time, I thought I could drive a car throught the national parks, I thought I could fall in love with someone. I had thought about so many good things for the first year of the new decades. But it haven't. At the end, I am still the "loser" in the lives, sitting in the room for the most of this year, doing something that no one really likes. Sometimes you just can't find out the places in this world, questioning if your decisions are good or not.
For many times, I fell into the illusion that I am the center of the world, and for many times when that illusion breaks, it hurts a lot. After all, I am just nobody. And my life is just sitting somewhere for the most of the day, seeing friends leaving but cannot do anything, seeing people are having fun while locking myself in the room.

Those happy scenes from 2019 still pop up in my mind often, these scenes are warm, but also like a cold damn sword, stabbing into my heart. For the uncoutable times when I have this feeling, it just stucks in my mind, nowhere to tell, but just cry and hid into the deepest place in my mind.

For the upcoming 2021, I have no brave to ask for bless again. But I do hope you can take care yourself and be well in the future, no matter when you saw this message.

Sorry, but I just don't know whom I should tell these to.